Elapsed time
Now displays the time elapsed while playing the track which is a close approximation to the time in the track.
Updated permissions on commands because this was stupid
Added more aliases for seeing the playing song
Also removed the copyright information due to copy and paste shenanigans; it'll probably return next update. Credit where credit is due.