Playlist support
Adds the playlist command, which adds all tracks from a YT playlist to the queue. Currently cannot take playlist data from a song in the playlist.
Started working on a command to add a song to the beginning of the queue, but this broke multiple times. Currently there's just a second copy of the play command commented out or something like that.
Fixed the elapsed timer.
Copyright is back.