Changed the member leave notification
Sometimes the text of the member leave notification displays the username of the person properly, sometimes the mention string fails to format correctly. This doesn't even matter because the embed title also contains the user's name, but it annoys me just enough to want to change it.

Just make sure I didn't do it wrong, ok?